+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add `allow_ip_edits` argument on `Wikidatum::Client.new`. This protects users from making IP address-exposing edits if they haven't explicitly opted-in to doing that. The argument defaults to false.
 - Add code to raise errors when various types of invalid input are passed to `Wikidatum::Client#add_statement`.
 
+### Fixed
+
+- Update the serializers and `add_statement` method to account for various changes that have been made upstream to the Wikibase REST API.
+
 ## 0.2.1 - 2022-08-13
 ### Fixed